requirements.txt
google-cloud-aiplatform[adk,agent-engines]==1.93.0
pydantic>=2.11.1,<3.0.0
python-dotenv==1.0.1
google-genai==1.16.1
google-adk==1.0.0
deprecated>=1.2.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0 